requirements.txt
streamlit
pandas
matplotlib
openpyxl
pyexcel
pyexcel-xls
numpy
fpdf
Pillow
opencv-python
